refactor(buildings): extract hall-connected suite packing + enclosure
Lift HallsToShafts → pack → door/walls into `HallConnectedGroups` /
`HallEnclosedSuites` so `LivableApartments` is only residential fill. Raise
default suite connectivity to 2.0 m and mildly prefer less-skinny grow cells.

## Agent Dialogue

- **Developer:** Extract general HTS+group pipeline; raise connectivity to ~2 m;
  lightly reduce narrow apartments; doors/walls are general too — only
  LivableApartment fill is bespoke.
- **Agent:** Added `hall_connected_suites` (pack + enclose layers), thinned
  `livable_apartments`, set `MIN_GROUP_CONNECTIVITY` to 2.0, aspect-aware grow
  score; updated CONTRIBUTING.
