+++ maybraid/richmond/buildings/CONTRIBUTING.md
@@ -106,23 +106,22 @@ and the per-interior galleries (`mini-mart-examples`, `public-restroom-examples`
    shared edges between primary rects.
 3. Full\* fills each primary rect with [`LivableApartments`](src/usage_areas/livable_apartments.rs):
    sample [`LivableApartmentsParameterized`](src/usage_areas/livable_apartments.rs)
-   (target m² catalog, Les Halles stall-door style) →
-   [`HallsToShafts`](src/usage_areas/halls_to_shafts.rs) → split residuals →
-   `pack_apartments_to_targets` → **one hall door per group** →
-   partition / hall-edge walls (no per-cell shells) →
-   [`LivableApartment`](src/usage_areas/livable_apartment/) layout
+   (target m² catalog) →
+   [`HallEnclosedSuites`](src/usage_areas/hall_connected_suites.rs)
+   ([`HallsToShafts`](src/usage_areas/halls_to_shafts.rs) → split →
+   [`pack_apartments_to_targets`](src/usage_areas/plan_cells.rs) with
+   [`MIN_GROUP_CONNECTIVITY`](src/usage_areas/plan_cells.rs) **2.0 m** →
+   one hall door / group + partition / hall-edge walls) →
+   [`LivableApartment`](src/usage_areas/livable_apartment/) fill per suite
    (entryway → [`RectPassageCluster`](src/usage_areas/rect_passage_cluster.rs)
-   max-rects + spanning-tree passages → per-rect
-   [`RectangularLivableArea`](src/usage_areas/rectangular_livable_area.rs) with
-   open/closed normalize). Shared XZ helpers live in
-   [`plan_geom`](src/usage_areas/plan_geom.rs); RLA carve sizes in
-   [`slot_policy`](src/usage_areas/rectangular_livable_area/slot_policy.rs).
-   Strategies default to `CaseAttempt` (AllOpen / SingleClosed ≤36 m² with one
-   port / SpineHall / GuillotineSplit). Min hall / access clear is **1.0 m**.
-   Apartment-level walls enclose bedrooms / bathrooms only; entry + common stay
-   open-plan (subtypes may wall themselves). Ungrouped / unfilled pockets stay
-   [`SpaceKind::InternalSpace`](src/fit.rs); Full\* maps those leftovers to
-   [`SpaceKind::ClosetSpace`](src/fit.rs).
+   → per-rect
+   [`RectangularLivableArea`](src/usage_areas/rectangular_livable_area.rs)).
+   Shared XZ helpers: [`plan_geom`](src/usage_areas/plan_geom.rs); RLA carve
+   sizes: [`slot_policy`](src/usage_areas/rectangular_livable_area/slot_policy.rs).
+   Strategies default to `CaseAttempt`. Min hall clear inside apartments is
+   **1.0 m**. Bedroom / bathroom walls only at apartment level; ungrouped /
+   soft-failed pockets stay [`SpaceKind::InternalSpace`](src/fit.rs); Full\*
+   maps leftovers to [`SpaceKind::ClosetSpace`](src/fit.rs).
 
 Playground: `/show i-apartment-floor-plan`, `/show i-apartment-floor-plan-examples`,
 `/show i-apartment-full-storey`, `/show i-apartment-full-storey-examples` (gallery),
