+++ maybraid/richmond/buildings/src/lib.rs
@@ -106,7 +106,9 @@ pub use usage_areas::{
 	CommercialStallStripParameterized, CommercialStallStripPlan, CommonBedroom,
 	CommonBedroomParameterized, CommonBedroomPlan, DiningRoom, DiningRoomParameterized,
 	DiningRoomPlan, EatingArea, EatingAreaParameterized, EatingAreaPlan, EnclosedRoom,
-	EnclosedRoomMins, EnclosedRoomParams, HallsToShafts, HallsToShaftsOptions, Janitorial, Kitchen,
+	EnclosedRoomMins, EnclosedRoomParams, HallConnectedGroups, HallEnclosedSuites,
+	HallSuiteEncloseParams, HallSuitePackParams, HallsToShafts, HallsToShaftsOptions, Janitorial,
+	Kitchen,
 	KitchenCounterLayout, KitchenParameterized, KitchenPlan, KnickKnackStall,
 	KnickKnackStallParameterized, KnickKnackStallPlan,
 	LivableApartment, LivableApartments, LivableApartmentsOptions, LivableApartmentsParameterized,
